Changed my.delay from int to float

+++ src/setup.h
@@ -162,7 +162,7 @@ struct CONFIG
   BOOLEAN display;      /* boolean, display the thread id verbose  */
   BOOLEAN config;       /* boolean, prints the configuration       */
   int     cusers;       /* default concurrent users value.         */
-  int     delay;        /* range for random time delay, see -d     */
+  float   delay;        /* range for random time delay, see -d     */
   int     timeout;      /* socket connection timeout value, def:10 */
   BOOLEAN bench;        /* signifies a benchmarking run, no delay  */
   BOOLEAN internet;     /* use random URL selection if TRUE        */
